pyproject.toml
[project]
name = "ai-agents-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "logfire>=3.17.0",
    "pydantic-ai>=0.2.14",
    "python-dotenv>=1.1.0",
]
